Refactor ScatterTile to use Double coords in [0,1)

+++ NoiseGeneration/ScatterTile.cs
@@ -6,14 +6,14 @@ namespace NoiseGeneration
 {
 	public struct ScatterTile
 	{
-		public List<int[]> m_points; //List of points populated by noise algorithm.
+		public List<double[]> m_points; //List of points populated by noise algorithm.
 		public int m_seed;
 		public int m_dims;
 
 		public ScatterTile(int seed, int dims){
 			m_dims = dims;
 			m_seed = seed;
-			m_points = new List<int[]> ();
+			m_points = new List<double[]> ();
 		}
 	}
 }
